# DistrictEnterpriseStatisticalData.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DistrictEnterpriseStatisticalData.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/enterprise": {
      "get": {
        "tags": [
          "Enterprise"
        ],
        "summary": "Получение всех предприятий",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnterpriseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnterpriseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnterpriseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Enterprise"
        ],
        "summary": "Создание предприятия",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Enterprise"
        ],
        "summary": "Обновление информации о предприятии",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/enterprise/{id}": {
      "get": {
        "tags": [
          "Enterprise"
        ],
        "summary": "Получение предприятия по регистрационному номеру",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Enterprise"
        ],
        "summary": "Удаление предприятия",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/enterprise-type": {
      "get": {
        "tags": [
          "EnterpriseType"
        ],
        "summary": "Получение всех типов предприятий",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnterpriseTypeDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnterpriseTypeDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnterpriseTypeDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EnterpriseType"
        ],
        "summary": "Создание типа предприятия",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseTypeCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseTypeCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseTypeCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EnterpriseType"
        ],
        "summary": "Обновление информации о типе предприятия",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseTypeDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseTypeDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EnterpriseTypeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/enterprise-type/{id}": {
      "get": {
        "tags": [
          "EnterpriseType"
        ],
        "summary": "Получение типа по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EnterpriseType"
        ],
        "summary": "Удаление типа предприятия",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnterpriseTypeDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/form-of-ownership": {
      "get": {
        "tags": [
          "FormOfOwnership"
        ],
        "summary": "Получение всех форм",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormOfOwnershipDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormOfOwnershipDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FormOfOwnershipDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FormOfOwnership"
        ],
        "summary": "Создание формы",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormOfOwnershipCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormOfOwnershipCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FormOfOwnershipCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FormOfOwnership"
        ],
        "summary": "Обновление информации о форме",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FormOfOwnershipDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FormOfOwnershipDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FormOfOwnershipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/form-of-ownership/{id}": {
      "get": {
        "tags": [
          "FormOfOwnership"
        ],
        "summary": "Получение формы по идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FormOfOwnership"
        ],
        "summary": "Удаление форме",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FormOfOwnershipDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EnterpriseCreateDto": {
        "required": [
          "address",
          "employees_number",
          "enterprise_type_id",
          "form_of_ownership_id",
          "name",
          "phone_number",
          "total_area"
        ],
        "type": "object",
        "properties": {
          "enterprise_type_id": {
            "type": "integer",
            "description": "Тип отрасли",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Наименование",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "description": "Телефон",
            "nullable": true
          },
          "form_of_ownership_id": {
            "type": "integer",
            "description": "Форма собственности",
            "format": "int32"
          },
          "employees_number": {
            "type": "integer",
            "description": "Количество работающих",
            "format": "int32"
          },
          "total_area": {
            "type": "integer",
            "description": "Общая площадь",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания предприятия"
      },
      "EnterpriseDto": {
        "required": [
          "address",
          "employees_number",
          "enterprise_type",
          "form_of_ownership",
          "name",
          "phone_number",
          "total_area"
        ],
        "type": "object",
        "properties": {
          "registration_number": {
            "type": "integer",
            "description": "Регистрационный номер",
            "format": "int32"
          },
          "enterprise_type": {
            "$ref": "#/components/schemas/EnterpriseTypeDto"
          },
          "name": {
            "type": "string",
            "description": "Наименование",
            "nullable": true
          },
          "address": {
            "type": "string",
            "description": "Адрес",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "description": "Телефон",
            "nullable": true
          },
          "form_of_ownership": {
            "$ref": "#/components/schemas/FormOfOwnershipDto"
          },
          "employees_number": {
            "type": "integer",
            "description": "Количество работающих",
            "format": "int32"
          },
          "total_area": {
            "type": "integer",
            "description": "Общая площадь",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "DTO предприятия"
      },
...
      "EnterpriseTypeCreateDto": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Тип предприятия",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания типа предприятия"
      },
      "EnterpriseTypeDto": {
        "required": [
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "description": "Тип предприятия",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO типа предприятия"
      },
...
      "FormOfOwnershipCreateDto": {
        "required": [
          "form"
        ],
        "type": "object",
        "properties": {
          "form": {
            "type": "string",
            "description": "Форма собственности",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO для создания формы собственности"
      },
      "FormOfOwnershipDto": {
        "required": [
          "form"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Идентификатор",
            "format": "int32"
          },
          "form": {
            "type": "string",
            "description": "Форма собственности",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "DTO формы собственности"
      },
...
    }
  }
}